chore(dependabot): Remove enable auto-merge
After some debugging and discussing with internal security squad, the core problem for the this permission issue with auto-merge is
```
GraphQL: ["Pull request User is not authorized for this protected branch"] (enablePullRequestAutoMerge)
```

The settings "Restrict who can push it to protected branch" rule we have `main` branch.

And currently no way to add `github-action` user to this list (there is an open discussion
https://github.com/orgs/community/discussions/25305

I proposed an idea to use some "real" user's (even grafana-bot) token to do the "auto-merge". But that is still under review by security
squad.

Meanwhile, to unblock merging any dependabot PR, I'm removing this check.

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>