Introduces cache to TSDB postings (#9621)
**What this PR does / why we need it**:
Introduces a new interface for fetching TSDB postings and implements a cached TSDB postings.
The new postings interface is named `PostingsReader` and contains two implementations: a simple one, which is identical to the former Postings behavior, and a new one, that caches postings using the existing `index_queries_cache_config`. By default, the simple/former implementation is used. The new cached implementation can be configured in the following way:
```yaml
storage_config:
  tsdb_shipper:
    cache_postings: true
```

In the future, we'll remove the `cache_postings` flag and change the behavior to cache postings by default.