serialize empty head block as well while checkpointing to retain the head block format (#10182)
**What this PR does / why we need it**:
When the head block is empty, we do not serialize it while
checkpointing. This causes a mismatch in chunk and head block format
which ideally should be on the same version. The problem happens when we
rollback from v4 chunk and block format to v3 and the head is empty in
checkpoint which then [gets set to the default
version](https://github.com/grafana/loki/blob/ce91076dd4cb6016f1f39b7a5126c2fd88a8b18e/pkg/ingester/checkpoint.go#L106)
i.e v3 while the checkpointed chunk would be v4.

**Checklist**
- [x] Tests updated

---------

Co-authored-by: Salva Corts <salva.corts@grafana.com>