bugfix/promtail.sh: Updated script to work better by default in k8s environment (#10390)
**What this PR does / why we need it**:

This PR fixes a bug in
[promtail.sh](https://github.com/grafana/loki/blob/94a5f94cf408e2da3e4293e8163e18c47c74e4d6/tools/promtail.sh).

promtail.sh is a script to deploy promtail in kubernetes clusters as a
daemonset. The defaults should make sense in that environment.

Currently, The default log parser in pipeline stages is, `- docker: `.
Kubernetes uses CRI format so it makes more sense to default to CRI
parser and this PR does exactly that.


This PR does not affect any existing users of loki and it'll improve it
for people who decide to use this script with the default arguments in
the future because kubernetes uses cri format and grafana can't parse
JSON logs properly if they were fed to loki with docker parser instead
of the more appropriate cri parser.