Add deprecation notes to Storage documentation (#10383)
**What this PR does / why we need it**:

Legacy storage backends (that are not single store aka "shipper") have been deprecated with https://github.com/grafana/loki/pull/10341 The docs should reflect this change.

**Special notes for your reviewer**:

The structure of the page changed _slightly_:
* Made object storage a separate section under "Chunk storage"
* Refer to this section in the "Single Store" section for supported backends
* Moved Azure example underneath the rest of the single store examples

While it is not yet perfect, I tried to change as little as possible and keep the big change for when the legacy storage backends are removed.

---------

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>
(cherry picked from commit 738ccbb372ad8d25d1638249226182db55072f75)