Distributor: add ingester append timeouts error (#10456)
**What this PR does / why we need it**:
Failing to send samples to ingesters because the request exceeded its
timeout is a very clear signal that ingesters are unable to keep up with
demand. In an incident today we saw that ingesters' push latencies were
increased sharply by an expensive regex query which was starving other
goroutines of time on CPU.

This new alert `loki_distributor_ingester_append_timeouts_total` will
give us a high-signal metric which we can use for alerting.