Remove `use_boltdb_shipper_as_backup` configuration
The setting `use_boltdb_shipper_as_backup` (`-tsdb.shipper.use-boltdb-shipper-as-backup`) was a remnant from the development of the TSDB storage.
It was used to allow writing to both TSDB and BoltDB when TSDB was still highly experimental.
Since TSDB is now stable and the recommended index type, the setting has become irrelevant and therefore was removed.
The previous default value `false` is applied.

Also closes https://github.com/grafana/loki/issues/9603

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>