Fix regression when parsing numbers in Push request (#10550)
**What this PR does / why we need it**:
Even though the[ Loki HTTP API docs for the push endpoint][1] state that
the stream label values should be strings, we previously didn't enforce
this requirement. With https://github.com/grafana/loki/pull/9694, we
started enforcing this requirement, and that broke some users.

In this PR we are reverting this type of assertion and adding a bunch of
tests to avoid the regression in the future.

[1]:
https://grafana.com/docs/loki/latest/reference/api/#push-log-entries-to-loki

(cherry picked from commit 5ab9515020658860053590c9e38e0262be78a9b1)