clients/pkg/promtail: tests no longer listen on 0.0.0.0
As a result, when you run the tests on a Mac, they no longer prompt
the user to ask if they are OK with the tests listening on an external
address. This makes the tests substantially easier to run on a Mac, in
addition to improving this library's security posture.

Updates #10407.