drone(docker-driver): Fetch tags before pushing image during release
We realized, loki-docker-driver image was not being pushed after 2.8.2. And looking into it's drone pipeline and comparing it with other steps, docker-driver step is not
fetching the tags, thus the final hash calculated to push is some HEAD branch.

e.g: even if we have tag 2.9.1, it pushes only with latest HEAD commit hash.

This PR fixes it.

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>