helm: Add complete object schema to single-binary volumeClaimTemplate to avoid synchronization issues (#10519)
**What this PR does / why we need it**:
This PR completes the object schema for the volumeClaim template inside
the single-binary helm chart. Currently it omits the apiVersion and the
kind.
This is working, because kubernetes auto-fills those values, however it
is causing an issue when you are trying to synchronize your helm chart
with your deployment (in my case with argoCD). It will be shown as out
of sync, because kubernetes has the fields, while the chart does not.
Syncing works, because kubernetes accepts it without apiVersion and
kind, but adds it itself again. That results in an endless sync-loop,
that could be avoided by adding these to the helm template.

Since both values are already filled in by kubernetes, this will not
break anything nor result in any diff for existing deployments.

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] If the change is worth mentioning in the release notes, add
`add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/setup/upgrade/_index.md`
- [x] For Helm chart changes bump the Helm chart version in
`production/helm/loki/Chart.yaml` and update
`production/helm/loki/CHANGELOG.md` and
`production/helm/loki/README.md`. [Example
PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)

Co-authored-by: Michel Hollands <42814411+MichelHollands@users.noreply.github.com>