Format + Library for building bloom filter "blocks" from chunks (#10780)
This PR introduces both a binary format for bloom filter "blocks" and a
library for building/querying them.

Each "block" consists of two files:
1) `Index`. This is a list of individually compressed pages of
stream+chunk information sorted in ascending fingerprint order. Each
series lists both the chunks indexed in the bloom filters as well as the
relevant offsets for said series in the bloom file.
2) `Blooms`. This is a list of individually compressed pages of bloom
filters for each series, again sorted in ascending fingerprint order.


There is still more work to do, but I'm opening this PR as an initial
pass.