fix(promtail): correctly parse list of drop stage sources from YAML (#10848)
**What this PR does / why we need it**:
This PR fixes YAML parsing of the `source` field in the drop stage when
said field is a list of strings. Because this list is not being
recognized in the code as a list of strings, but rather as a list of
generic `interface{}` elements, it fails a type check and throws an
error. In order to fix this, an `interface{}` list is manually converted
to a `string` list. The problem was already reproducible with existing
tests, but the test YAML mistakenly referred to the field as `sources`
rather than `source`.

**Which issue(s) this PR fixes**:
Fixes #10095 

**Special notes for your reviewer**:

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] If the change is worth mentioning in the release notes, add
`add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/setup/upgrade/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in
`production/helm/loki/Chart.yaml` and update
`production/helm/loki/CHANGELOG.md` and
`production/helm/loki/README.md`. [Example
PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)

---------

Co-authored-by: Michel Hollands <42814411+MichelHollands@users.noreply.github.com>