Adjust bloom gateway request structure
Since we already group chunkRefs by fingerprint on the client side to
determine the correct bloom gateway to send the request to, we can keep
the grouping for the gRPC request.
This will not only optimize the request size, but also helps to reduce
complexity of querying blooms on the bloom gateway itself.

There is still conversion between []*logproto.ShortRef and v1.ChunkRefs
required, which could be removed by sending v1.ChunkRefs over the wire.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>