chore(renovate): make sure to `go mod tidy` after upgrade
This avoids issues like the ones in this PR (make check-mod failure)
https://drone.grafana.net/grafana/loki/29104/3/10

Core issue seems to be old entries not removed from `go.sum` file so `go mod tidy` updates it and CI not happy with the diff.

This PR fixes it. Tested in on my fork.
https://github.com/kavirajk/loki/commit/bece62796adc9b64f2dd780db8849a7ea89cf6d2

PR that correctly updates go.sum
https://github.com/kavirajk/loki/pull/33/files

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>