Add basic structure of bloom compactor (#10748)
This pull request adds the basic structure for the new bloom compactor component.

- Adds new `bloom-compactor` target that runs with with multiple instances joined by a ring
- Adds boilerplate functions to index blooms and compact blocks

The main goal of this PR is to provide a basic functionality on which future smaller PRs can build upon. Since the code path is completely separate in a new component, it is not used anywhere at the moment.