Deprecated/Deleted Config Checker tool (#10977)
**What this PR does / why we need it**:

This PR adds a new _deprecated/deleted config checker_ tool to allow
users upgrading from previous Loki versions to check their Loki config
and runtime config for deprecated and deleted options, as well as
providing them with alternatives and reasons why those configs were
deprecated/deleted.

Mimir provides a [tool][1] that allows users to migrate their old
configs to new ones. Having that is our future goal but the new tool in
this PR is a good middle ground for the short-mid term.

We also update the checklist in the PR template to ask contributors to
update the tool source of truth if the contributor is deprecating or
deleting a config.

Here's an example for how to use the tool:
```bash
go run tools/deprecated-config-checker/main.go \
  -config.file tools/deprecated-config-checker/test-fixtures/config.yaml \
  -runtime-config.file tools/deprecated-config-checker/test-fixtures/runtime-config.yaml
```

And this is an example output:

![image](https://github.com/grafana/loki/assets/8354290/a86c0815-9875-4086-a267-9bbe06a6ee0e)

### Notes

I think I've gone through all the deprecated and deleted configs. I
cannot find the config pointed by the deprecated `ruler.storage`. It
states
```
ruler:
    ...
    # Deprecated: Use -ruler-storage. CLI flags and their respective YAML config
    # options instead. 
    storage:
          ...
```

I couldn't find that `-ruler-storage` cli flag nor the YAML definition.

[1]:
https://grafana.com/docs/mimir/latest/manage/tools/mimirtool/#convert