always align volume range timestamp to end of step (#11136)
**What this PR does / why we need it**:
We had originally assumed that it would be weird for a user to get back
a timeseries response from `volume_range` that didn't include a
datapoint with a timestamp at the start of their requested range. This
was a bad assumption. The `volume_range` endpoint is inherently doing
aggregations. For each step, a datapoint is calculated representing the
percent of chunk volumes that the selector appeared in. It therefore
stands to reason that the steps volume was not that value until the end
of the step time range, as we are to assume volume is monotonically
increasing. The fact that we were using the start of the range made it
difficult to turn this data into a per-second rate, which this PR aims
to fix.

**Which issue(s) this PR fixes**:
Fixes #11134