Enable/disable bloom gateway per tenant
This commits adds a per-tenant setting to enable/disable the use of the
bloom gateway component for filtering chunks.

As an operator, you want to control what tenant can use the filtering of
chunks based on search string, because the creation of the bloom filters
requires a significant amount of resources.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>