Propagate trace ID with HTTP gRPC request. (#11251)
The changes in https://github.com/grafana/loki/pull/10688 did not
propage the trace ID from the context. `Frontend.RoundTripGRPC` would
inject the trace ID into the request. That's not done in `Frontend.Do`.
This changes extends the `codec.EncodeRequest` to inject the trace ID
there. This is more inline with other metadata.