Bloom tokenization performance tweaks (#11338)
**What this PR does / why we need it**:

The main work of this PR is to improve the performance of the
reassemble() function. By changing the assignment of the 'cur' variable
to not do addition, and incrementing the 'pos' variable separately, the
benchmarks improved as follows:

BenchmarkTokens/threeSkip1Chunk/v2-10  	   88864	     13487 ns/op
BenchmarkTokens/threeSkip1ChunkAt2/v2At2-10 98607 12256 ns/op

(This was validated with multiple runs, on multiple size inputs)

The remaining changes in the PR are just to add some benchmarking for
PopulateSeriesWithBloom, in case we need to do other comparisons for
future improvements.