Bloom Gateway: process blocks immediately when they are available (#11360)
At current point in time, the worker calls
`w.store.GetBlockQueriersForBlockRefs()` to obtain a list of block
queriers sorted by fingerprint in ascending order.
However, the shipper used by the store to fetch blocks by its refs,
calls a callback function when a requested block is available. Instead
of waiting for all callbacks to be called and then returning a single
list of block queriers, the store now exposes a new function `ForEach()`
that also accepts a callback that is passed to the underlying shipper.

This allows to process blocks in the worker as soon as they become
available. Some block may be served from cache, some may need to be
downloaded. That way, we reduce waiting time.

---------

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>