Add trivy filesystem scanning (#11462)
**What this PR does / why we need it**:

Adds trivy filesystem scanning to our Makefile and PR pipelines. This
will scan the `go.mod` file for vulnerabilities in dependencies, which
are sometimes missed by just scanning the image.

PS @dannykopping this also adds the changes to the vulnerabilities
output we talked about that should make it easier to identify a) where
the vulnerability is (ie. docker image / os dep vs go.mod dep), and b)
what version the dependency needs to be updated to in order to fix it.