Run compaction over existing bloom blocks (#11486)
**What this PR does / why we need it**:
Upon changes to TSDB indexes, the existing blooms needs to be amended
with newly added chunks, or new series should be added to blocks.
This PR compares the existing meta information with the most up to date
TSDB information in the job. In case of no changes to indexes,
compaction is skipped; if there's any block needs amending, it does that
and merges all blocks in fp range; if there's no previous compaction, it
compacts from scratch.

**Special notes for your reviewer**:
This logic does not support cutting blocks upon topology changes to
bloom-compactors. It can create blocks with series outside of the fp
range of the compactor. Cutting blocks will be addressed in a follow-up
PR.