Fix index gateway startup if bloom gateway is enabled
The index gateways discovers the bloom gateways by using the bloom
gateway ring.
In order to initialize the ring before the index gateway, it needs to be
added as dependency of the index gateway.

The dependency is conditional, because we don't want the ring to be
initialized even though the bloom gateway is not used.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>