Bloomshipper: Use `model.Time` in `MetaRef` and `BlockRef`
The `Ref` of `MetaRef` and `BlockRef` do have start and end timestamps,
which are of type `int64`.
This however, can lead to the problem that different components might
use different precisions (`s`, `ms`, `ns`) for these timestamps.

This PR changes the current implementation of the bloom shipper to use
`ms` precision timestamps (using `model.Time`) instead of `s`, to match
the behaviour that we have to encode ChunkRefs to object store keys.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>